applied new folder structure
