+++ src/app/(auth)/login/layout.tsx
@@ -1,0 +1,9 @@
+import { type Metadata } from "next";
+export const metadata: Metadata = {
+  title: "Login",
+};
+const LoginLayout = ({ children }: { children: React.ReactNode }) => {
+  return <>{children}</>;
+};
+
+export default LoginLayout;
