+++ src/app/components/Footer/styles.ts
@@ -1,0 +1,29 @@
+import { createStyles, rem } from "@mantine/core";
+export const useStyles = createStyles((theme) => ({
+  footer: {
+    borderTop: `${rem(1)} solid ${
+      theme.colorScheme === "dark" ? theme.colors.dark[5] : theme.colors.gray[2]
+    }`,
+    backgroundColor:
+      theme.colorScheme === "dark" ? theme.colors.blue[3] : theme.white,
+    color: theme.colorScheme === "dark" ? theme.colors.dark[0] : theme.black,
+  },
+
+  inner: {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center",
+    paddingTop: theme.spacing.xl,
+    paddingBottom: theme.spacing.xl,
+
+    [theme.fn.smallerThan("xs")]: {
+      flexDirection: "column",
+    },
+  },
+
+  links: {
+    [theme.fn.smallerThan("xs")]: {
+      marginTop: theme.spacing.md,
+    },
+  },
+}));
