+++ src/app/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Footer } from "./Footer";
+export { default as Header } from "./Header";
+export { MainWrapper } from "./MainWrapper";
+export { ThemeSwitch } from "./ThemeSwitch";
