+++ src/app/favlist/page.tsx
@@ -1,25 +1,12 @@
-import Link from "next/link";
-import { type Metadata } from "next";
-import Button from "~/clients/components/Button";
-import Text from "~/clients/components/Text";
-import Card from "~/clients/components/FavCards";
-
-export const metadata: Metadata = {
-  title: "Map",
-  description: "Generated by create-t3-app",
-};
+"use client";
+import { Text } from "@mantine/core";
+import { Card } from "./compoments";
+import { useStyles } from "./styles";
 
 const FavPage = () => {
+  const { classes } = useStyles();
   return (
-    <div
-      style={{
-        display: "flex",
-        justifyContent: "center",
-        alignItems: "center",
-        flexDirection: "column",
-        marginTop: "10%",
-      }}
-    >
+    <div className={classes.container}>
       <Text fz="lg" fw={700} mb={20}>
         Your Favorite Tech Events
       </Text>
