+++ src/app/favlist/styles.ts
@@ -1,0 +1,10 @@
+import { createStyles } from "@mantine/core";
+export const useStyles = createStyles(() => ({
+  container: {
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+    flexDirection: "column",
+    paddingTop: "10%",
+  },
+}));
