--- src/app/signup/layout.tsx
@@ -1,9 +1,0 @@
-const SingupLayout = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <main className="flex h-screen flex-col items-center justify-center">
-      {children}
-    </main>
-  );
-};
-
-export default SingupLayout;
