--- src/clients/components/Header.tsx
@@ -1,156 +1,0 @@
-"use client";
-
-import { usePathname } from "next/navigation";
-import Link from "next/link";
-import Image from "next/image";
-import {
-  createStyles,
-  Header as MantineHeader,
-  Container,
-  Group,
-  Burger,
-  Paper,
-  Transition,
-  rem,
-} from "@mantine/core";
-import { useDisclosure } from "@mantine/hooks";
-import ThemeSwitch from "./ThemeSwitch";
-
-const HEADER_HEIGHT = rem(60);
-const LINKS = [
-  { link: "/", label: "Home" },
-  { link: "/map", label: "Map" },
-];
-
-const useStyles = createStyles((theme) => ({
-  root: {
-    position: "relative",
-    zIndex: 1,
-    backgroundColor:
-      theme.colorScheme === "dark" ? theme.colors.blue[3] : theme.white,
-    color: theme.colorScheme === "dark" ? theme.colors.dark[0] : theme.black,
-  },
-
-  dropdown: {
-    position: "absolute",
-    top: HEADER_HEIGHT,
-    left: 0,
-    right: 0,
-    zIndex: 0,
-    borderTopRightRadius: 0,
-    borderTopLeftRadius: 0,
-    borderTopWidth: 0,
-    overflow: "hidden",
-
-    [theme.fn.largerThan("sm")]: {
-      display: "none",
-    },
-  },
-
-  header: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    height: "100%",
-  },
-
-  links: {
-    [theme.fn.smallerThan("sm")]: {
-      display: "none",
-    },
-  },
-
-  burger: {
-    [theme.fn.largerThan("sm")]: {
-      display: "none",
-    },
-  },
-
-  link: {
-    display: "block",
-    lineHeight: 1,
-    padding: `${rem(8)} ${rem(12)}`,
-    borderRadius: theme.radius.sm,
-    textDecoration: "none",
-    color:
-      theme.colorScheme === "dark"
-        ? theme.colors.dark[0]
-        : theme.colors.gray[7],
-    fontSize: theme.fontSizes.sm,
-    fontWeight: 500,
-
-    "&:hover": {
-      backgroundColor:
-        theme.colorScheme === "dark"
-          ? theme.colors.dark[6]
-          : theme.colors.gray[0],
-    },
-
-    [theme.fn.smallerThan("sm")]: {
-      borderRadius: 0,
-      padding: theme.spacing.md,
-    },
-  },
-
-  linkActive: {
-    "&, &:hover": {
-      backgroundColor: theme.fn.variant({
-        variant: "light",
-        color: theme.primaryColor,
-      }).background,
-      color: theme.fn.variant({ variant: "light", color: theme.primaryColor })
-        .color,
-    },
-  },
-}));
-
-const Header = () => {
-  const activePathname = usePathname();
-  const [opened, { toggle }] = useDisclosure(false);
-  const { classes, cx } = useStyles();
-
-  const items = LINKS.map((link) => (
-    <Link
-      key={link.label}
-      href={link.link}
-      className={cx(classes.link, {
-        [classes.linkActive]: activePathname === link.link,
-      })}
-    >
-      {link.label}
-    </Link>
-  ));
-
-  return (
-    <MantineHeader height={HEADER_HEIGHT} className={classes.root}>
-      <Container className={classes.header}>
-        <div>
-          <Image src="/images/logo.png" alt="logo" width={45.5} height={57.5} />{" "}
-          Tech Meets
-        </div>
-        <Group spacing={5} className={classes.links}>
-          {items}
-          <ThemeSwitch />
-        </Group>
-
-        <Burger
-          opened={opened}
-          onClick={toggle}
-          className={classes.burger}
-          size="sm"
-        />
-
-        <Transition transition="pop-top-right" duration={200} mounted={opened}>
-          {(styles) => (
-            <Paper className={classes.dropdown} withBorder style={styles}>
-              {items}
-              <ThemeSwitch />
-            </Paper>
-          )}
-        </Transition>
-      </Container>
-    </MantineHeader>
-  );
-};
-
-export default Header;
