--- src/clients/providers/TrpcClient.tsx
@@ -1,61 +1,0 @@
-"use client";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { httpBatchLink, loggerLink } from "@trpc/client";
-import { createTRPCReact } from "@trpc/react-query";
-import { useState } from "react";
-import superjson from "superjson";
-import type { AppRouter } from "~/server/api/routers/root";
-import { env } from "~/env.mjs";
-
-export const trpc = createTRPCReact<AppRouter>({
-  unstable_overrides: {
-    useMutation: {
-      async onSuccess(opts) {
-        await opts.originalFn();
-        await opts.queryClient.invalidateQueries();
-      },
-    },
-  },
-});
-
-function getBaseUrl() {
-  if (typeof window !== "undefined")
-    // browser should use relative path
-    return "";
-  if (env.NEXT_PUBLIC_NETLIFY_URL)
-    // reference for vercel.com
-    return `https://${env.NEXT_PUBLIC_NETLIFY_URL }`;
-  if (process.env.RENDER_INTERNAL_HOSTNAME)
-    // reference for render.com
-    return `http://${env.NEXT_PUBLIC_LOCALHOST }:${
-      env.NEXT_PUBLIC_PORT 
-    }`;
-  // assume localhost
-  return `http://localhost:${(env.NEXT_PUBLIC_PORT ) ?? 3000}`;
-}
-
-function TrpcProvider(props: { children: React.ReactNode }) {
-  const [queryClient] = useState(() => new QueryClient());
-  const [trpcClient] = useState(() =>
-    trpc.createClient({
-      links: [
-        loggerLink({
-          enabled: () => true,
-        }),
-        httpBatchLink({
-          url: `${getBaseUrl()}/api/trpc`,
-        }),
-      ],
-      transformer: superjson,
-    })
-  );
-  return (
-    <trpc.Provider client={trpcClient} queryClient={queryClient}>
-      <QueryClientProvider client={queryClient}>
-        {props.children}
-      </QueryClientProvider>
-    </trpc.Provider>
-  );
-}
-
-export default TrpcProvider;
