+++ src/utils/index.ts
@@ -1,2 +1,4 @@
 export { mapTheme } from "./googleMapsTheme";
 export { loader } from "./googleMapsLoader";
+export { default as convertDate } from "./date-converter";
+export { default as theme } from "./theme";
